reafctor: merge product view and edit sidebars
